infra/flask-app/requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
prometheus-client==0.21.1
